fix(migrations/02-sales-migrations.js): :bug: Alteração feita para mudar valor do atributo "type" de "FLOAT" para "DECIMAL" no atributo "total_price"

+++ back-end/src/database/migrations/02-sales-migrations.js
@@ -30,7 +30,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(9,2),
       },
       delivery_address: {
         allowNull: false,
